+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,10 @@
 
 ## Next Release
 
+## 0.1.15
+
+- \[contracts\] Add BSC deployment data
+- \[router\] Dashboard update
 ## 0.1.15-beta.21
 
 - \[router\] Fix auto-rebalancing bug